requirements.txt
Django
psycopg2
psycopg2-binary
djangorestframework
requests
markdown==2.6.11
